+++ rpc/src/client/transport/router.rs
@@ -105,7 +105,7 @@ impl SubscriptionRouter {
     }
 
     /// Removes all the subscriptions relating to the given query.
-    pub fn remove_by_query(&mut self, query: impl ToString) -> usize {
+    pub fn remove_by_query(&mut self, query: impl ToString) -> u32 {
         self.subscriptions
             .remove(&query.to_string())
             .map(|subs_for_query| subs_for_query.len())
@@ -116,7 +116,7 @@ impl SubscriptionRouter {
 #[cfg(feature = "websocket-client")]
 impl SubscriptionRouter {
     /// Returns the number of active subscriptions for the given query.
-    pub fn num_subscriptions_for_query(&self, query: impl ToString) -> usize {
+    pub fn num_subscriptions_for_query(&self, query: impl ToString) -> u32 {
         self.subscriptions
             .get(&query.to_string())
             .map(|subs_for_query| subs_for_query.len())
